refactor file structure and add modifications in validation
